Advanced reading functions and plotting style
The reading functions handle now the "#" operator marking the extrapolated values. The only manual modification of the original lists now is the removal of the space in the atomic mass number (easy with Sublime text: cmd+A, cmd+shift+L) in the AME16 list.

Additionally, the plotting style was unified and uses now the color scheme of my dissertation. Now there is a clear distinction between measured and extrapolated values (thanks to the new reading functions).